# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Quan Li Lich Phong Lab",
        "version": "1.0.0",
        "title": "Lab Calender Server"
    },
    "schemes": {
        "http": "http",
        "https": "https"
    },
    "securityDefinitions": {
        "bearerAuth": {
            "name": "Authorization",
            "in": "header",
            "type": "apiKey",
            "description": "Bearer {token}"
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/user/getUserById/{uid}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Tìm user với mã UID",
                "description": "Trả về 1 người dùng và thông tin của người dùng",
                "operationId": "getUserById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Mã uid của user",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Invalid UID supplied"
                    },
                    "404": {
                        "description": "USER NOT FOUND"
                    }
                }
            }
        },
...
        "api/user/delete/{uid}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Xóa người dùng",
                "description": "Chỉ admin mới có thể xóa người dùng",
                "operationId": "deleteUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Mã uid của user",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Invalid UID supplied"
                    },
                    "404": {
                        "description": "USER NOT FOUND"
                    }
                }
            }
        },
...
    },
...
}